% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roams_SSM.R
\name{attach_insample_info}
\alias{attach_insample_info}
\title{Attach In-Sample Information to Fitted State Space Model}
\usage{
attach_insample_info(model)
}
\arguments{
\item{model}{A fitted model object of class \code{roams_SSM}, \code{classical_SSM}, \code{oracle_SSM}, \code{huber_robust_SSM}, or \code{trimmed_robust_SSM}.}
}
\value{
A modified version of the input model object, with an additional class \code{insample_info}, and the following in-sample elements appended:
\describe{
  \item{\code{filtered_states}}{Filtered state estimates using data up to each time point.}
  \item{\code{predicted_states}}{One-step-ahead state predictions.}
  \item{\code{filtered_observations}}{Expected observations given data up to each time point.}
  \item{\code{predicted_observations}}{One-step-ahead forecasts of observations.}
  \item{\code{filtered_states_var}}{List of filtered state variance matrices.}
  \item{\code{predicted_states_var}}{List of one-step-ahead state prediction variances.}
  \item{\code{predicted_observations_var}}{List of one-step-ahead observation forecast variances.}
  \item{\code{mahalanobis_residuals}}{Vector of Mahalanobis distances of residuals from predicted observations.}
}

For models of class \code{roams_SSM}, \code{classical_SSM}, or \code{oracle_SSM}, the following additional elements are also attached:
\describe{
  \item{\code{smoothed_states}}{Posterior means of hidden states using all data.}
  \item{\code{smoothed_observations}}{Posterior mean of the observed series based on smoothed states.}
  \item{\code{smoothed_states_var}}{List of smoothed state variance matrices.}
}

These smoothed attributes are obtained using the RTS smoothing algorithm  (Rauch et al. 1965).
}
\description{
Attaches detailed in-sample information—such as predicted, filtered, and smoothed states and observations—to a model object fitted using any of the package’s supported SSM estimation methods.
These quantities are not stored by default in model objects due to their potentially large memory footprint.
}
\details{
The attached outputs enable richer diagnostics, outlier inspection, and plotting.
For \code{huber_robust_SSM} and \code{trimmed_robust_SSM} models, in-sample information is computed using a custom robust filtering function, and smoothed quantities (\code{smoothed_states}, \code{smoothed_observations}, and \code{smoothed_states_var}) are \strong{not available}.
This function should only be applied once to a model object.
}
\references{
Rauch, H.E., Tung, F., Striebel, C.T. (1965). Maximum likelihood estimates of linear dynamic systems. \emph{AIAA Journal 3}(8), 1445–1450. https://doi.org/10.2514/3.3166
}
\seealso{
\code{\link{oos_filter}}
}
