% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{autoplot}
\alias{autoplot}
\title{A Cat Function}
\usage{
autoplot(model_list, attribute = "BIC")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
