% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model.R
\name{best_prop_outlying_model}
\alias{best_prop_outlying_model}
\title{A Cat Function}
\usage{
best_prop_outlying_model(model_list, target)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
