% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{compute_objective}
\alias{compute_objective}
\title{A Cat Function}
\usage{
compute_objective(y, par, build, gamma = array(0, dim(y)))
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
