% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{compute_objective_no_gamma}
\alias{compute_objective_no_gamma}
\title{A Cat Function}
\usage{
compute_objective_no_gamma(y, par, build, outlier_locs)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
