% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_methods.R
\name{delete_outliers_oracle_SSM}
\alias{delete_outliers_oracle_SSM}
\title{A Cat Function}
\usage{
delete_outliers_oracle_SSM(
  y,
  init_par,
  build,
  outlier_locs,
  lower = NA,
  upper = NA
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
