% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_methods.R
\name{gamma_start_SSM}
\alias{gamma_start_SSM}
\title{A Cat Function}
\usage{
gamma_start_SSM(
  y,
  lambda,
  init_par,
  build,
  gamma = array(0, dim(y)),
  B = 20,
  lower = NA,
  upper = NA
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
