% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{huber_robust_filter}
\alias{huber_robust_filter}
\title{A Cat Function}
\usage{
huber_robust_filter(y, par, build)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
