% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_methods.R
\name{huber_robust_method}
\alias{huber_robust_method}
\title{#' A Cat Function}
\usage{
huber_robust_method(y, init_par, build, lower = NA, upper = NA)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
