% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_methods.R
\name{huber_robust_SSM}
\alias{huber_robust_SSM}
\title{Huber-Robust State Space Model Fit}
\usage{
huber_robust_SSM(
  y,
  init_par,
  build,
  lower = NA,
  upper = NA,
  control = list(parscale = init_par)
)
}
\arguments{
\item{y}{A numeric matrix of observations (time points in rows).}

\item{init_par}{A numeric vector of initial parameter values.}

\item{build}{A function that returns a \code{dlm} model given a parameter vector. The \code{specify_SSM()} function can be used to create this \code{build} function.}

\item{lower}{Optional numeric vector of lower bounds for parameter estimation. Defaults to \code{-Inf}. Must be of same length as \code{init_par}.}

\item{upper}{Optional numeric vector of upper bounds for parameter estimation. Defaults to \code{Inf}. Must be of same length as \code{init_par}.}

\item{control}{Optional list of control parameters passed to \code{optim}. Default is \code{list(parscale = init_par)}, which can help the optimizer if parameters are on vastly different scales.}
}
\value{
An object of class \code{huber_robust_SSM} containing the optimization result, the original data, and the original build function.
}
\description{
Fits a robust state space model by minimizing a Huber loss objective as per Crevits and Croux (2018), providing protection against moderate outliers. The predicted observations used in the Huber loss are computed using the Huber robust filter from Cipra and Romera (1997).
}
\references{
Crevits R. and Croux C. (2018). Robust Estimation of Linear State Space Models. *Communications in Statistics: Simulation and Computation*

Cipra, T., Romera, R. (1997). Kalman filter with outliers and missing observations. *Test* 6, 379–395. https://doi.org/10.1007/BF02564705
}
\seealso{
\code{\link{trimmed_robust_SSM}}, \code{\link{roams_SSM}}, \code{\link[stats]{optim}}, \code{\link{attach_insample_info}}, \code{\link{oos_filter}}, \code{\link{specify_SSM}}
}
