% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robularized_SSM.R
\name{IPOD_oos_robust_filter}
\alias{IPOD_oos_robust_filter}
\title{A Cat Function}
\usage{
IPOD_oos_robust_filter(y, par, build, lambda)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
