% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robularized_SSM.R
\name{robularized_SSM}
\alias{robularized_SSM}
\title{A Cat Function}
\usage{
robularized_SSM(
  y,
  init_par,
  build,
  num_lambdas = 10,
  lowest_lambda = 2,
  highest_lambda = NA,
  num_lambdas_crowding = 0,
  cores = 1,
  B = 20,
  lower = NA,
  upper = NA,
  control = list()
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
