% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robularized_SSM_changing_ISV.R
\name{robularized_SSM_isv}
\alias{robularized_SSM_isv}
\title{A Cat Function}
\usage{
robularized_SSM_isv(
  y,
  init_par,
  build,
  num_lambdas = 10,
  lowest_lambda = 2,
  highest_lambda = NA,
  num_lambdas_crowding = 0,
  cores = 1,
  B = 20,
  lower = NA,
  upper = NA
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
