% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robularized_SSM.R
\name{run_IPOD}
\alias{run_IPOD}
\title{A Cat Function}
\usage{
run_IPOD(y, lambda, init_par, build, B, lower, upper, control)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
