% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_methods.R
\name{classical_SSM}
\alias{classical_SSM}
\title{A Cat Function}
\usage{
classical_SSM(y, init_par, build, lower = NA, upper = NA, control = list())
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
