% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roams_SSM.R
\name{oos_filter}
\alias{oos_filter}
\title{Compute Out-of-Sample Inference for Fitted State Space Model}
\usage{
oos_filter(
  y_oos,
  model,
  build,
  outlier_locs = rep(0, nrow(y_oos)),
  threshold = sqrt(qchisq(0.99, ncol(y_oos))),
  multiplier = 2
)
}
\arguments{
\item{y_oos}{A numeric matrix containing out-of-sample observations. Each row corresponds to a time point.}

\item{model}{A fitted model object of class \code{roams_SSM}, \code{classical_SSM}, \code{oracle_SSM}, \code{huber_robust_SSM}, or \code{trimmed_robust_SSM}.}

\item{build}{A function that maps a numeric parameter vector to a corresponding \code{dlm} model object. The \code{specify_SSM} function can be used to create this \code{build} function.}

\item{outlier_locs}{A logical or binary vector of the same length as \code{nrow(y)}, indicating time points to be treated as missing (i.e., time points that are known to be outliers). Used only with \code{oracle_SSM} models.}

\item{threshold}{Mahalanobis distance threshold for identifying outliers in \code{roams_SSM} models. Default is \code{sqrt(qchisq(0.99, ncol(y)))}.}

\item{multiplier}{Multiplier for how quickly the filter grows its filtered state variance (uncertainty) after detecting an outlier in \code{roams_SSM} models. Default is \code{2}.}
}
\value{
A named list containing out-of-sample inference results:
\describe{
  \item{\code{filtered_states}}{Filtered state estimates using out-of-sample data.}
  \item{\code{predicted_states}}{One-step-ahead state predictions.}
  \item{\code{filtered_observations}}{Expected observations given past out-of-sample data.}
  \item{\code{predicted_observations}}{One-step-ahead forecasts of observations.}
  \item{\code{filtered_states_var}}{List of filtered state variance matrices.}
  \item{\code{predicted_states_var}}{List of one-step-ahead state prediction variances.}
  \item{\code{predicted_observations_var}}{List of one-step-ahead observation forecast variances.}
  \item{\code{mahalanobis_residuals}}{Vector of Mahalanobis distances of residuals from predicted observations.}
  \item{\code{outliers_flagged}}{Vector of 1's and 0's indicating whether timepoints are flagged as outlying or not based on the \code{threshold} supplied (only available if \code{model} is of class \code{roams_SSM}).}
}
}
\description{
Applies the fitted model parameters to a user-supplied out-of-sample dataset to compute predicted and filtered states and observations. Robust and classical inference procedures are supported depending on the class of the input model.
}
\details{
The function reuses the model's fitted parameters to generate inference on new data \code{y_oos}. Robust variants use appropriate robust filters, while the classical and oracle models use standard Kalman filtering. For \code{oracle_SSM} models, observations flagged in \code{outlier_locs} are treated as missing during filtering.
}
\seealso{
\code{\link{attach_insample_info}}, \code{\link{specify_SSM}}
}
