% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_methods.R
\name{oracle_SSM}
\alias{oracle_SSM}
\title{A Cat Function}
\usage{
oracle_SSM(y, init_par, build, outlier_locs, B = 20, lower = NA, upper = NA)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
