% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model.R
\name{outlier_target_model}
\alias{outlier_target_model}
\title{Select Model Based on Target Outlier Proportion}
\usage{
outlier_target_model(model_list, target)
}
\arguments{
\item{model_list}{An object of class \code{robularized_SSM_list}.}

\item{target}{A numeric value between 0 and 1 indicating the desired proportion of outlying observations.}
}
\value{
A single \code{robularized_SSM} object whose estimated outlier proportion is closest to \code{target}.
}
\description{
Extracts the model from a \code{robularized_SSM_list} object whose estimated outlier proportion
is closest to a user-specified target.
}
\examples{
\dontrun{
# Select the model with an outlier proportion closest to 10\%
target_model <- outlier_target_model(models, target = 0.1)
}

}
