% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_SSM.R
\name{specify_SSM}
\alias{specify_SSM}
\title{A Cat Function}
\usage{
specify_SSM(
  state_transition_matrix,
  state_noise_var,
  observation_matrix,
  observation_noise_var,
  init_state_mean,
  init_state_var
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
