% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_methods.R
\name{trimmed_robust_SSM}
\alias{trimmed_robust_SSM}
\title{#' A Cat Function}
\usage{
trimmed_robust_SSM(
  y,
  init_par,
  build,
  alpha,
  lower = NA,
  upper = NA,
  control = list()
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\value{
description
}
\description{
This function allows you to express your love of cats.
}
\details{
Additional details...
}
